+++ mundy/mesh/tests/performance_tests/PerfTestNgpAggregates.cpp
@@ -18,6 +18,7 @@
 // **********************************************************************************************************************
 // @HEADER
 
+
 #define ANKERL_NANOBENCH_IMPLEMENT
 
 // C++ core libs
@@ -43,7 +44,7 @@
 #include <stk_mesh/base/Selector.hpp>
 
 // Mundy libs
-#include <mundy_math/Vector3.hpp>      // for mundy::math::Vector3
+#include <mundy_math/Vector3.hpp>      // for mundy::Vector3
 #include <mundy_mesh/Aggregate.hpp>    // for mundy::mesh::Aggregate
 #include <mundy_mesh/BulkData.hpp>     // for mundy::mesh::BulkData
 #include <mundy_mesh/MeshBuilder.hpp>  // for mundy::mesh::MeshBuilder
@@ -180,7 +181,7 @@ void run_test() {
     bulk_data.declare_relation(elem, node, 0);
 
     // Populate the fields
-    vector3_field_data(node_center_field, node).set(1.1 * i, 2.2 * i, 3.3);
+    vector3_field_data(node_center_field, node).set(1.1 * static_cast<double>(i), 2.2 * static_cast<double>(i), 3.3);
     vector3_field_data(node_force_field, node).set(5.0, 6.0, 7.0);
     vector3_field_data(node_velocity_field, node).set(1.0, 2.0, 3.0);
     scalar_field_data(elem_radius_field, elem).set(0.5);
